+++ common/src/main/java/de/maxhenkel/voicechat/gui/VoiceChatScreenBase.java
@@ -39,7 +39,11 @@ public abstract class VoiceChatScreenBase extends Screen {
 
     @Override
     public void render(GuiGraphics guiGraphics, int mouseX, int mouseY, float delta) {
-        renderTransparentBackground(guiGraphics);
+        if (isIngame()) {
+            renderTransparentBackground(guiGraphics);
+        } else {
+            renderDirtBackground(guiGraphics);
+        }
         renderBackground(guiGraphics, mouseX, mouseY, delta);
         super.render(guiGraphics, mouseX, mouseY, delta);
         renderForeground(guiGraphics, mouseX, mouseY, delta);